Reflection, Inspector rework. applstdefJc.rework, Simulink platform files * Fwc/objectBaseC.j gets more capability: CLASS_C ObjectArrayJc, reflection_Types, CLASS_C Int32ARRAY moved from ObjectJc * FBaccessNode: new registerNode_AccessNode_Inspc(...) used for registration of a the Simulink FB with this name. * REFJc is now defined with a simple reference (not enhanced) in some applstdefJc.h,   therefore rework some usages of references. Don't use .ref immediately, always use REFJc, TODO regard in Java2C-translation. * ReflectionJc: searchTrc_ClassJc moved to start of file, it wasn't conspicuous, necessary for debug faulty reflection generation. * ReflectionJc: getString_FieldJc(...) now returns some more useable info: address of field and content. Usefull for debugging. * ApplstdefJc.h: Some templates or systems now in 'incApplSpecific' subdir. * rtwtypes.h etc. as part of platform specific files in CRuntimeJavalike   (able to use instead generated files from simulink, with necessary adaptions for CRuntimeJavalike types).

+++ source/Fwc/fw_threadContext.c
@@ -114,7 +114,7 @@ MemC getUserBuffer_ThreadContextFw(int size, char const* sign, ThreadContextFW_s
             _thCxt->addrFree += size;
             _thCxt->bitAddrUsed |= 1 << ix;
             _thCxt->addrUsed[ix].sign = sign;
-            _thCxt->ixLastAddrUsed = ix;
+            _thCxt->ixLastAddrUsed = (int16)ix;
             set_MemC(_thCxt->addrUsed[ix].used, addr,size);  //store the found range.
             init0_MemC(ret);
             return ret; //NOTE: the user is responsible for saving its content.
